+++ docs/changelog.rst
@@ -8,25 +8,37 @@ v0.1
 * Data structures ``static_truth_table`` and ``dynamic_truth_table``
   `#1 <https://github.com/msoeken/kitty/pull/1>`_
 
-* Bit-access functions ``set_bit`` and ``get_bit``
+* Bit functions: ``set_bit``, ``get_bit``, ``clear_bit``, ``clear``
   `#1 <https://github.com/msoeken/kitty/pull/1>`_
+  `#8 <https://github.com/msoeken/kitty/pull/8>`_
 
 * Constructors ``create_nth_var``, ``create_from_binary_string``, ``create_from_hex_string``, and ``create_majority``
   `#1 <https://github.com/msoeken/kitty/pull/1>`_
   `#4 <https://github.com/msoeken/kitty/pull/4>`_
   `#5 <https://github.com/msoeken/kitty/pull/5>`_
 
-* Unary and binary operations: ``unary_not``, ``binary_and``, ``binary_or``, and ``binary_xor``
+* Unary and binary operations: ``unary_not``, ``unary_not_if``, ``binary_and``, ``binary_or``, and ``binary_xor``
   `#2 <https://github.com/msoeken/kitty/pull/2>`_
+  `#8 <https://github.com/msoeken/kitty/pull/8>`_
 
 * Ternary operations: ``ternary_majority`` and ``ternary_ite``
   `#3 <https://github.com/msoeken/kitty/pull/3>`_
 
-* Binary predicates: ``equal``
+* Binary predicates: ``equal``, ``less_than``
   `#4 <https://github.com/msoeken/kitty/pull/4>`_
+  `#8 <https://github.com/msoeken/kitty/pull/8>`_
+
+* Operators: ``~``, ``==``, ``<``
+  `#8 <https://github.com/msoeken/kitty/pull/8>`_
 
 * Swap adjacent variables: ``swap_adjacent_inplace``, ``swap_adjacent``
   `#6 <https://github.com/msoeken/kitty/pull/6>`_
 
+* Swap variables: ``swap_inplace``, ``swap``
+  `#8 <https://github.com/msoeken/kitty/pull/8>`_
+
 * Flip variable: ``flip_inplace``, ``flip``
   `#7 <https://github.com/msoeken/kitty/pull/7>`_
+
+* NPN canonization: ``exact_npn_canonization``, ``create_from_npn_config``
+  `#8 <https://github.com/msoeken/kitty/pull/8>`_